DTOs e Data Annotations

+++ Backend/src/ProEventos.API/Controllers/EventosController.cs
@@ -3,7 +3,7 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using ProEventos.Application.Contratos;
-using ProEventos.Domain;
+using ProEventos.Application.Dtos;
 
 namespace ProEventos.API.Controllers
 {
@@ -24,8 +24,8 @@ namespace ProEventos.API.Controllers
                try
                {
                     var eventos = await _eventoService.GetAllEventosAsync(true);
-                    if (eventos == null) return NotFound("Nenhum evento encontrado.");
-
+                    if (eventos == null) return NoContent();
+                    
                     return Ok(eventos);
                }
                catch (Exception ex)
@@ -41,7 +41,7 @@ namespace ProEventos.API.Controllers
                try
                {
                     var evento = await _eventoService.GetEventoByIdAsync(id, true);
-                    if (evento == null) return NotFound("Evento por Id não encontrado");
+                    if (evento == null) return NoContent();
 
                     return Ok(evento);
                }
@@ -58,7 +58,7 @@ namespace ProEventos.API.Controllers
                try
                {
                     var eventos = await _eventoService.GetAllEventosByTemaAsync(tema, true);
-                    if (eventos == null) return NotFound("Eventos por tema não encontrados.");
+                    if (eventos == null) return NoContent();
 
                     return Ok(eventos);
                }
@@ -70,7 +70,7 @@ namespace ProEventos.API.Controllers
           }
 
           [HttpPost]
-          public async Task<IActionResult> Post(Evento model)
+          public async Task<IActionResult> Post(EventoDto model)
           {
                try
                {
@@ -86,7 +86,7 @@ namespace ProEventos.API.Controllers
                }
           }
           [HttpPut("{id}")]
-          public async Task<IActionResult> Put(int id, Evento model)
+          public async Task<IActionResult> Put(int id, EventoDto model)
           {
                try
                {
@@ -106,14 +106,17 @@ namespace ProEventos.API.Controllers
           {
                try
                {
+                    var eventos = await _eventoService.GetAllEventosAsync(true);
+                    if (eventos == null) return NoContent();
+
                     return await _eventoService.Delete(id) ?
                        Ok("Evento Deletado com sucesso!") :
-                       BadRequest("Erro ao tentar deletar evento!");
+                       throw new Exception("Aconteceu algum erro ao tentar deletar este evento");
                }
                catch (Exception ex)
                {
                     return this.StatusCode(StatusCodes.Status500InternalServerError,
-                    $"Erro ao tentar recuperar eventos. Erro {ex.Message}");
+                    $"Erro ao tentar deletar eventos. Erro {ex.Message}");
                }
           }
